Add id's to packets to ease troubleshooting. Use uint32_t for payload length. Add packet_stream API to process packets in sequence while doing multithreading. Rename packet_copy() into packet_clone() and add flags to prevent copying the buffer. Fix small memory leak in input_dvb. Add packet_stream support to proto_mpeg_ts. Print packet processing result in core_process_dump_pkt_info(). Add locks in timer API. Add queue reference to timers to dequeue them more easily and faster.

+++ include/pom-ng/proto.h
@@ -53,7 +53,7 @@ struct proto_dependency {
 struct proto_process_stack {
 	struct proto_reg *proto;
 	void *pload;
-	size_t plen;
+	uint32_t plen;
 
 	struct packet_info *pkt_info;
